Move the Route invocation from Application to the Service implementation file to avoid SBT's macro recompilation issue.
