+++ server/src/main/scala/controllers/Application.scala
@@ -1,21 +1,15 @@
 package controllers
 
-import java.nio.ByteBuffer
-
-import boopickle.Default._
 import play.api.mvc._
-import services.ApiService
+import services.ApiServer
+import services.ApiServer._
 import spatutorial.shared.Api
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
-object Router extends autowire.Server[ByteBuffer, Pickler, Pickler] {
-  override def read[R: Pickler](p: ByteBuffer) = Unpickle[R].fromBytes(p)
-  override def write[R: Pickler](r: R) = Pickle.intoBytes(r)
-}
 
 object Application extends Controller {
-  val apiService = new ApiService()
+
 
   def index = Action {
     Ok(views.html.index("SPA tutorial"))
@@ -29,13 +23,8 @@ object Application extends Controller {
       val b = request.body.asBytes(parse.UNLIMITED).get
 
       // call Autowire route
-      Router.route[Api](apiService)(
-        autowire.Core.Request(path.split("/"), Unpickle[Map[String, ByteBuffer]].fromBytes(ByteBuffer.wrap(b)))
-      ).map(buffer => {
-        val data = Array.ofDim[Byte](buffer.remaining())
-        buffer.get(data)
-        Ok(data)
-      })
+      ApiServer.route(path, b).map(Ok(_))
+
   }
 
   def logging = Action(parse.anyContent) {
